+++ EnemyEncounterTool/src/ThugLevel2.java
@@ -1,8 +1,7 @@
 public class ThugLevel2 extends ThugLevel1 implements  WeaponUser{
     public boolean doesLeapAttack;
     public int leapAttackCooldown;
-
-    public ThugLevel2(String name, int health, int walkSpeed, int runSpeed, int viewDistance, int blindedMovementSpeed, String description, int attackDamage, int initialAttackTimer, int minAttackTimer, int maxAttackTimer, boolean canUseWeapon, ComboDatabase comboDB, boolean regularCombos, boolean doesLeapAttack, int leapAttackCooldown) {
+    public ThugLevel2(String name, int health, double walkSpeed, double runSpeed, double viewDistance, double blindedMovementSpeed, String description, int attackDamage, int initialAttackTimer, int minAttackTimer, int maxAttackTimer, boolean canUseWeapon, ComboDatabase comboDB, boolean regularCombos, boolean doesLeapAttack, int leapAttackCooldown) {
         super(name, health, walkSpeed, runSpeed, viewDistance, blindedMovementSpeed, description, attackDamage, initialAttackTimer, minAttackTimer, maxAttackTimer, canUseWeapon, comboDB, regularCombos);
         this.doesLeapAttack = doesLeapAttack;
         this.leapAttackCooldown = leapAttackCooldown;
